// src/server/model/pLedger.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface PLedger { 
    id?: number;
    name?: string;
    printName?: string;
    ledgerGroupId?: number;
    ledgerGroupName?: string;
    openingBal?: number;
    gstin?: string;
    stateName?: string;
    stateId?: number;
    mobile?: string;
    email?: string;
    cityId?: number;
    cityName?: string;
    bankAcNo?: string;
    bankName?: string;
    branchName?: string;
    bankIfscCode?: string;
    alternateNo1?: string;
    alternateNo2?: string;
    faxNo?: string;
    panNumber?: string;
    gstType?: PLedger.GstTypeEnum;
    systemGenerated?: boolean;
}
export namespace PLedger {
    export type GstTypeEnum = 'Regular' | 'Composition' | 'Unregistered';
    export const GstTypeEnum = {
        Regular: 'Regular' as GstTypeEnum,
        Composition: 'Composition' as GstTypeEnum,
        Unregistered: 'Unregistered' as GstTypeEnum
    };
}